requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
black==19.10b0
bleach==4.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
conda-pack==0.6.0
cycler==0.11.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
flake8==3.9.2
fonttools==4.28.5
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==2.6.0
grpcio==1.43.0
hydra-colorlog==1.1.0
hydra-core==1.1.0
idna==3.3
imageio-ffmpeg==0.4.3
imageio==2.10.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipykernel==6.4.1
ipython-genutils==0.2.0
ipython==7.27.0
jedi==0.18.0
jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
markdown==3.3.6
markupsafe==2.0.1
matplotlib-inline==0.1.2
matplotlib==3.5.1
mccabe==0.6.1
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
ninja==1.10.2.2
notebook==6.4.5
numpy==1.21.3
oauthlib==3.2.0
omegaconf==2.1.1
opensimplex==0.4.2
packaging==21.0
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
pip==21.0.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydeprecate==0.3.1
pyflakes==2.3.1
pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyspng==0.1.0
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
pyzmq==22.2.1
regex==2021.11.2
requests-oauthlib==1.3.1
requests==2.26.0
rich==11.2.0
rsa==4.8
scikit-image==0.17.2
scipy==1.7.1
send2trash==1.8.0
sentry-sdk==1.5.6
setuptools==59.5.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
tifffile==2021.11.2
toml==0.10.2
torch-tb-profiler==0.3.1
torch==1.7.1+cu110
torchmetrics==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.10
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.3
wheel==0.37.0
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0